+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 
 Refactoring of dialogs, drawers and menus, using the `Phoenix.LiveView.JS` API - and largely based on `CoreComponent`'s modal component. These changes reduce reliance on additional JavaScript, improve alignment with standard practice, and include accessibility improvements.
 
-See `PrimerLive.Component.dialog/1` for details.
+See component documentation for further details:
+- [Action menu](`PrimerLive.Component.action_menu/1`)
+- [Dialog](`PrimerLive.Component.dialog/1`)
+- [Drawer](`PrimerLive.Component.drawer/1`)
+- [Dropdown](`PrimerLive.Component.dropdown/1`)
+- [Select menu](`PrimerLive.Component.select_menu/1`)
 
 ### Additions
 
@@ -53,7 +58,6 @@ phx-click={open_dialog("my-dialog")}
 phx-click={close_dialog("my-dialog")}
 ```
 
-- All dialogs, drawers and menus need to have an id attribute.
 - Form state: the previous method to preserve state, using "a fictitious and unique field name" can be removed.
 - Because `focus_first` (without a selector) is now the default, nothing needs to be changed when using this attribute. If in existing code a selector value is used, rename the attribute to `focus_after_opening_selector`.
 - Replace `prompt_options` and `phx_click_touch` with `status_callback_selector`. There's no simple way to replace `prompt_options`, because passing JavaScript functions is no longer supported. A solution coiuld be very similar to the previous `phx_click_touch` method. This example LiveComponent may give some ideas how to approach the update: